CamelCase component names and move tests to src/.
